refactor all layers.

+++ src/main/java/com/melnikov/bulish/my/budget/my_budget_backend/model/DeletePurchaseNotification.java
@@ -2,9 +2,8 @@ package com.melnikov.bulish.my.budget.my_budget_backend.model;
 
 import com.melnikov.bulish.my.budget.my_budget_backend.enums.NotificationStatus;
 
-public class DeletePurchaseNotification extends AbstractNotification<Integer> {
-
-    public DeletePurchaseNotification(Integer entityID) {
+public class DeletePurchaseNotification extends AbstractNotification<Long> {
+    public DeletePurchaseNotification(Long entityID) {
         super("PURCHASE", NotificationStatus.DELETED, entityID);
     }
 }
